:recycle: Removed SDGSettingsMixin and fixed locations
SDGSettingsMixin added a setting and the org_type_cfg which was already
added in the context processeor anyway.
The core template for some reason contained the if-statement to decide to
show products or not based on a setting. This was moved to the product
specific template, where it is actually used.